[cleanup][xl] Code refractored and cleanup
