+++ ckanext/fcscopendata/logic/action/__init__.py
@@ -1,0 +1,25 @@
+from ckanext.fcscopendata.logic.action.get import (
+    package_search,
+    package_show,
+    organization_show,
+    group_show,
+    tag_show,
+    vocabulary_show,
+)
+
+from ckanext.fcscopendata.logic.action.create import (
+    package_create,
+    resource_create,
+    organization_create,
+    group_create,
+    tag_create,
+    vocabulary_create,
+)
+
+from ckanext.fcscopendata.logic.action.update import (
+    package_update,
+    resource_update,
+    organization_update,
+    group_update,
+    vocabulary_update,
+)
